Split HTTP Actions into separate partial classes.
