+++ src/HttpClientServiceHelper/Delete.cs
@@ -1,0 +1,23 @@
+﻿using System;
+using System.Net.Http;
+using System.Threading.Tasks;
+
+namespace HttpClientServiceHelper
+{
+    public partial class HttpClientHelper
+    {
+        /// <summary>
+        /// Triggers a DELETE request to the specified route and retrieves the result as a string which you can serialize.
+        /// </summary>
+        /// <param name="Route"></param>
+        /// <returns></returns>
+        public static async Task<HttpResponseMessage> DeleteAsync(string Route)
+        {
+            using (HttpClient httpClient = new HttpClient())
+            {
+                var httpResponse = await httpClient.DeleteAsync(new Uri(Route));
+                return httpResponse;
+            }
+        }
+    }
+}
