+++ src/HttpClientServiceHelper/HttpClientServiceHelper.csproj
@@ -12,8 +12,8 @@ It returns an HttpResponseMessage/String Response depending on the method called
     <PackageLicenseFile>LICENSE</PackageLicenseFile>
     <GeneratePackageOnBuild>false</GeneratePackageOnBuild>
     <RepositoryUrl>https://github.com/DaraOladapo/HttpClientServiceHelper</RepositoryUrl>
-    <ReleaseVersion>0.2</ReleaseVersion>
-    <Version>1.0.1</Version>
+    <ReleaseVersion>0.3</ReleaseVersion>
+    <Version>1.0.3</Version>
     <PackageRequireLicenseAcceptance>true</PackageRequireLicenseAcceptance>
   </PropertyGroup>
